refactor: split main.py
