+++ server/mahjong_game/table.py
@@ -1,8 +1,8 @@
 import random
 from typing import List
 
-from tile import Tile
-from wall import Wall
+from mahjong_game.tile import Tile
+from mahjong_game.wall import Wall
 
 
 class Table:
