+++ server/mahjong_game/utils.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from tile import Tile
+from mahjong_game.tile import Tile
 
 
 def sort_tiles_by_id(tiles: List[Tile]):
